+++ src/main/java/Lesson_2/Homework/SelectionSort.java
@@ -1,20 +1,34 @@
-package Lesson_2.Homework.sort;
+package Lesson_2.Homework;
 
 public class SelectionSort {
     public static void sort(Integer[] arr) {
-        for (int i = 0; i < arr.length - 1; i++) {
-            int min = i;
+        int limit = arr.length;
 
-            for (int j = min + 1; j < arr.length; j++) {
+        for (int i = 0; i < limit - 1; i++) {
+            int min = i;
+            int max = i;
+
+            for (int j = min + 1; j < limit; j++) {
                 if (arr[j] < arr[min]) {
                     min = j;
                 }
+
+                if (arr[j] > arr[max]) {
+                    max = j;
+                }
             }
 
-            int temp = arr[i];
-            arr[i] = arr[min];
-            arr[min] = temp;
-
+            swapArrayElements(arr, i, min);
+            if (i != max) {
+                swapArrayElements(arr, limit - 1, max);
+                limit--;
+            }
         }
+    }
+
+    public static void swapArrayElements (Integer[] arr, int i, int j) {
+        int temp = arr[i];
+        arr[i] = arr[j];
+        arr[j] = temp;
     }
 }
